♻️ refactoring : Mypage -> Profile
